Fix bug and Finish UI User

+++ webapp/src/main/java/com/controller/GroupController.java
@@ -8,6 +8,7 @@ import java.time.LocalDate;
 import java.time.ZoneId;
 import java.time.format.DateTimeFormatter;
 import java.io.IOException;
+import java.io.PrintWriter;
 //import java.sql.Date;
 import java.util.List;
 import java.util.Date;
@@ -168,8 +169,17 @@ public class GroupController {
     @RequestMapping(value = "/deleteGroup",method = RequestMethod.POST)
 	@Transactional(rollbackOn = { Exception.class })
 	public void deleteGroup(HttpServletRequest request,HttpServletResponse response) throws IOException {
-		 System.out.println("i'm svc");
-	     groupService.deleteGroupDetail(request);
+    	Integer groupId = Integer.parseInt(request.getParameter("groupId"));
+        if (!registrationRepository.findStudentsByGroupId(groupId).isEmpty()) {
+            // Display dialog
+            response.setContentType("text/html");
+            PrintWriter out = response.getWriter();
+            out.println("<script type=\"text/javascript\">");
+            out.println("alert('This Group was registrated by Student !Can't Delete ');");
+            out.println("</script>");
+        } else {
+            groupService.deleteGroupDetail(request);
+        }
 	}
